% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsmm_emission_functions.R
\name{available_marginal_emission_dist}
\alias{available_marginal_emission_dist}
\title{Provides the list of currently supported marginal emission distributions and the list of their parameters.}
\usage{
available_marginal_emission_dist()
}
\value{
A \code{data.frame} with the following columns:
\code{type} (continuous, discrete or categorical), \code{distribution}, \code{params}, \code{parameter_type} and \code{parameter_size}
}
\description{
Provides the list of currently supported marginal emission distributions and the list of their parameters.
}
\examples{
available_marginal_emission_dist()
}
\seealso{
\code{available_marginal_emission_viz_options()} for the list of available visualization options for each of these distributions.
}
