% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsmm_sojourn_functions.R
\name{available_sojourn_dist}
\alias{available_sojourn_dist}
\title{Provides the list of currently supported sojourn distributions and the list of their parameters.}
\usage{
available_sojourn_dist()
}
\value{
A data.frame with the following columns: \code{distribution_type} and \code{parameters}
}
\description{
Provides the list of currently supported sojourn distributions and the list of their parameters.
}
\examples{
available_sojourn_dist()
}
