% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsmm_functions.R
\name{.compute_obs_probs}
\alias{.compute_obs_probs}
\title{Computes matrix b for a given sequence of observations.}
\usage{
.compute_obs_probs(model, X)
}
\description{
Computes matrix b for a given sequence of observations.
}
