% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsmm_sojourn_functions.R
\name{gammafit}
\alias{gammafit}
\title{Estimates gamma distribution parameters using method of moments}
\usage{
gammafit(x, wt = NULL)
}
\arguments{
\item{x}{a vector of values that need to be described by a Gamma distribution.}

\item{wt}{(optional) a vector of weights with values in [0,1] and of the same length of \code{x}.}
}
\description{
Estimates gamma distribution parameters using method of moments
}
