% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsmm_functions.R
\name{initialize_hsmm}
\alias{initialize_hsmm}
\title{Initialize a hidden semi-Markov model.}
\usage{
initialize_hsmm(model, nmin = NULL, verbose = FALSE, seq_sim_seed = NULL)
}
\arguments{
\item{model}{a \code{hsmm_spec} object. A specified hidden semi-Markov model. Use the function \code{specify_hsmm()} to specify a hidden semi-Markov model.}

\item{nmin}{(optional) a strictly positive integer. XXXX}

\item{verbose}{(optional) a logical. If TRUE, the function prints the internal steps of the function.}

\item{seq_sim_seed}{(optional) an integer. XXX}
}
\value{
an object of class \code{hsmm}, which can be used to decode observation sequences with the function \code{predict_states_hsmm()} or which can be fitted to observations with the function \code{fit_hsmm()}.
}
\description{
This function initializes the joint emission probabilities given the states,
i.e. determines the values of \eqn{Pr(X_i, E_i | S_j)}  where \eqn{X_i} are the observations at a given time-point \eqn{i}, \eqn{E_i} are the values of the augmented variables at time-point \eqn{i} and \eqn{S_j} is the state \eqn{j}.
}
\examples{
my_model_spec = simple_model_spec
class(my_model_spec)
my_model_init = initialize_hsmm(my_model_spec)
class(my_model_init)

}
