% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_hsmm_em_par}
\alias{plot_hsmm_em_par}
\title{Visualize the marginal emission distribution of a \code{hsmm} or \code{hsmm_spec} model.}
\usage{
plot_hsmm_em_par(model, verbose = FALSE, show_all_missing_probs = TRUE)
}
\arguments{
\item{model}{a \code{hsmm} or \code{hsmm_spec} object specifying the model for which the marginal emission distribution should be visualized.}

\item{verbose}{(optional) a logical specifying if the internal steps of the function should be printed.}
}
\value{
a ggplot object.
}
\description{
Visualize the marginal emission distribution of a \code{hsmm} or \code{hsmm_spec} model.
}
