% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_hsmm_fit_param}
\alias{plot_hsmm_fit_param}
\title{Visualizes the status of the EM-procedure.}
\usage{
plot_hsmm_fit_param(model, title = NULL)
}
\arguments{
\item{model}{the output of the \code{fit_hsmm()} function.}

\item{title}{(optional) an character specifying the title of the visualization (typically the name of the model). By default, there is no title.}
}
\value{
a ggplot object.
}
\description{
If a model has been fitted using the function \code{fit_hsmm()}, this function can be used to visualize the convergence of the EM.
}
