% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_hsmm_marg_dist}
\alias{plot_hsmm_marg_dist}
\title{Visualizes the marginal emission distribution of a \code{hsmm} model.}
\usage{
plot_hsmm_marg_dist(model, show_missing_probs = TRUE, verbose = FALSE)
}
\arguments{
\item{model}{a \code{hsmm} object specifying the model for which the marginal emission distributions should be visualized.}

\item{verbose}{(optional) a \code{logical} specifying if the internal steps of the function should be printed.}

\item{show_missing_prob}{(optional) a \code{logical} specifying if transparency should be used to reflect how likely variables are going to be missing in each state.}
}
\value{
a ggplot object.
}
\description{
Visualizes the marginal emission distribution of a \code{hsmm} model.
}
\examples{
my_model = simple_model
plot_hsmm_marg_dist(model = my_model)
}
