% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_hsmm_seq}
\alias{plot_hsmm_seq}
\title{Plots a single sequence of observation.}
\usage{
plot_hsmm_seq(X, model, verbose = FALSE, selection = data.frame())
}
\arguments{
\item{X}{a \code{data.frame} specifying the sequence of observation. Should contains data for only one sequence. Each column is a variable.
Any column starting with the characters \code{state} is considered to be a state sequence and will be displayed on top of the observations.
If two columns start with \code{state}, then a third line will display the aggrement between these two columns. This is useful if one desires to compare a decoded sequence with the ground truth or state sequence resulting from the decoding of models with different parameters.}

\item{model}{a \code{hsmm} or \code{hsmm_spec} object specifying the model associated with the observation sequence.}

\item{verbose}{(optional) a logical specifying if the internal steps of the function should be printed.}
}
\value{
a ggplot object.
}
\description{
Plots a single sequence of observation.
}
\examples{
my_model = my_model_spec
Xsim = simulate_hsmm(model = my_model, n_state_transition = 20)
plot_hsmm_seq(X = Xsim, model = my_model)

}
