% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_hsmm_sojourn_dist}
\alias{plot_hsmm_sojourn_dist}
\title{Visualizes the sojourn distributions of a hidden semi-Markov model.}
\usage{
plot_hsmm_sojourn_dist(model, maxt = 100)
}
\arguments{
\item{model}{a \code{hsmm} or \code{hsmm_spec} object specifying the model for which the sojourn distributions should be visualized.}

\item{maxt}{(optional) an integer specifying the upper limit for the x-axis, which is the sojourn length.}
}
\value{
a ggplot object.
}
\description{
Visualizes the sojourn distributions of a hidden semi-Markov model.
}
