% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_hsmm_state_colors}
\alias{plot_hsmm_state_colors}
\title{Visualization of the model's state colors.}
\usage{
plot_hsmm_state_colors(model)
}
\arguments{
\item{model}{a \code{hsmm} model.}
}
\description{
Visualization of the model's state colors.
}
