% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_hsmm_transitions}
\alias{plot_hsmm_transitions}
\title{Visualizes the model graph with transition probabilities.}
\usage{
plot_hsmm_transitions(
  model,
  size = 20,
  label_size = NULL,
  label_color = "white",
  arrow_gap = NULL
)
}
\arguments{
\item{model}{a \code{hsmm} or \code{hsmm_spec} object specifying the model for which the transition probabilities should be visualized.}

\item{size}{(optional) the size of the nodes (hidden states).}

\item{label_size}{(optional) the size of the labels, i.e. hidden state names.}

\item{label_color}{(optional) the color of the labels. Default is white. Any valid color specification can be used here.}

\item{arrow_gap}{(optional) a \code{double} value in [0,1] specifying the gap between the arrow tips and the nodes.
If \code{NULL} (default value), the function adapts the value of this parameter to the specified size.}
}
\value{
a ggplot object.
}
\description{
The edges width is proportional to the transition probability.
}
