% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsmm_functions.R
\name{predict_states_hsmm}
\alias{predict_states_hsmm}
\title{Predicts the hidden state sequence from observations}
\usage{
predict_states_hsmm(
  model,
  X,
  method = "viterbi",
  ground_truth = data.frame(),
  trust_in_ground_truth = 0.75,
  verbose = FALSE,
  graphical = FALSE,
  c_code = "original",
  ...
)
}
\arguments{
\item{model}{a \code{hsmm} object. The model used to predict the hidden sequence of states.}

\item{X}{a \code{data.frame}. The observations.}

\item{method}{a \code{character} specifying the decoding algorithm.
\code{method = "viterbi"} returns the most likely sequence of hidden states.
\code{method = "smoothed"} applies the "Forward-Backward" algorithm as described by Guédon, 2003, and returns the "smoothed" (posterior) probability of each state at each time-point.}

\item{ground_truth}{(optional) a \code{data.frame}.}
}
\description{
This function predict the most likely hidden states from observations.
Two methods are implemented:
the first one applies the Viterbi algorithm and predicts the most likely sequence of hidden states,
the second one applies the Forward-Backward algorithm and returns the probabilitiy of each state at each time-point.
}
\examples{
my_model_spec = simple_model_spec
Xsim = simulate_hsmm(my_model_spec, n_state_transitions = 20)
my_model_init = initialize_hsmm(my_model_spec)
viterbi = predict_states_hsmm(model = my_model_init, X = Xsim, method = "viterbi")
Xsim$state_viterbi = viterbi$state_seq$state
plot_hsmm_seq(X = Xsim, model = my_model_init)

}
