% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simple_model}
\alias{simple_model}
\title{A simple hidden semi-Markov model for function demonstration purposes.}
\format{
An object of class \code{hsmm} of length 10.
}
\usage{
simple_model
}
\description{
A \code{hsmm} object specified as a 2-state 3-variables model.
}
\seealso{
This model is the same as the one specified in the example presented for \code{specify_hsmm()}.
Type \code{?specify_hsmm} to read the specification code.
}
\keyword{datasets}
