% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsmm_functions_simulate.R
\name{simulate_hsmm}
\alias{simulate_hsmm}
\title{Simulate a state and observation sequence from a hidden semi-Markov model.}
\usage{
simulate_hsmm(
  model,
  seq_id = NULL,
  n_state_transitions = NULL,
  n_timepoints = NULL,
  seed = NULL,
  all_states = FALSE,
  min_tp_per_state = NULL,
  mult_seq_allowed = TRUE,
  state_seq = NULL
)
}
\arguments{
\item{model}{a object of class \code{hsmm_spec} or \code{hsmm}.}

\item{seq_id}{(optional) a character specifying the name of the sequence to simulate.}

\item{n_state_transitions}{(optional) XXX}

\item{n_timepoints}{(optional) XXX}

\item{seed}{(optional) XXX}

\item{all_states}{(optional) XXX}

\item{min_tp_per_state}{(optional) XXX}

\item{mult_seq_allowed}{(optional) XXX}
}
\value{
A data.frame with the following columns: TODO xxxx
}
\description{
This function returns a \code{data.frame} with a simulated state sequence and observations.
The length of the simulated sequence can be specified either as a number of state transition (with argument \code{n_state_transitions}) or as a number of time-point (with argument \code{n_timepoints})
Each row of the returned \code{data.frame} is a time-point and the columns specify the sequence id, the time-point, the state and the values of the observation. Each variable has its own column.
}
\examples{
my_model_spec = specify_hsmm(
   J = 2,
   init = c(1,0),
   transition = matrix(c(0,1,1,0),2,2),
   sojourn = list(type = "gamma", shape = c(2,10), scale = c(10,3)),
   parms.emission = list(
       var1 = list(
           type = "norm",
           params = list(
               mean = c(0,1),
               sd = c(0.3,0.2)
               ),
           missing_prob = c(0.6,0.1)
           ),
     var2 = list(
         type = "binom",
         params = list(
             size = rep(1,2),
             prob = c(0.2,0.8)
             ),
         missing_prob = c(0,0)
         ),
     var3 = list(
         type = "non-par",
         params = list(
             values = c("a","b","c","d"),
             probs = matrix(c(0.7,0.1,0.1,0.1,
                            1/4,1/4,1/4,1/4), 4,2)
           ),
         missing_prob = c(0.5,0.5)
         )
     ),
   state_names = c("A","B"),
   state_colors = c("seagreen1","lightcoral")
)
class(my_model_spec)
Xsim = simulate_hsmm(model = my_model_spec)
plot_hsmm_seq(model = my_model_spec, X = Xsim)
}
\keyword{HSMM}
