% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsmm_functions.R
\name{specify_hsmm}
\alias{specify_hsmm}
\title{Specifies a hidden semi-Markov model}
\usage{
specify_hsmm(
  J,
  init,
  transition,
  sojourn,
  parms.emission,
  state_names = NULL,
  state_colors = NULL,
  augment_data_fun = NULL
)
}
\arguments{
\item{J}{an integer. The number of hidden states.}

\item{init}{a vector of double. The initial probabilities associated with each states. Must be a vector of length \code{J}. If values do not sum to 1 a warning message is displayed and the vector is divided by its sum.}

\item{transition}{a matrix of double. The transition probabilities.
Must be a matrix of size \code{J x J} in which the element \code{(r, c)} provides the probability of transitioning from state \code{r} to state \code{c}.
The diagonal elements must be equal to zero except for absorbing states. For absorbing states, all the other elements of that row must be zero. For other states, if there are non-zero diagonal elements, a warning message is displayed and the elements are set to zero.
If the sum over the rows is different that one, a warning message is displayed and the rows are divided by their sum.}

\item{sojourn}{a list. The sojourn distributions. The list must have at least two elements: a \code{type} element which specifies the type of sojourn distributions and the other elements are the distribution parameters.
Use the function \code{available_sojourn_dist()} to see the supported sojourn distributions.}

\item{parms.emission}{a list. The marginal emission distributions. TODO: explain how to format them.}

\item{state_names}{(optional) a vector of characters. Names associated to each state. Must be of length \code{J}. If unspecified, the states are numbered from 1 to \code{J}}

\item{state_colors}{(optional) a vector of color-specifying characters. Colors associated to each state. Must be of length \code{J}. If unspecified, the colors are picked from the \code{viridis} palette.}

\item{augment_data_fun}{(optional) a function that takes as argument a matrix of observation (see xxx TODO for how observations must be formatted) and returns a matrix of augmented variables. Additionally, this function must have xxx TODO xxx explain the requirements.}
}
\value{
An object of class \code{hsmm_spec} which can be used to simulate data following the model specification and which can be initialized with the function \code{initialize_hsmm()}.
}
\description{
This function returns an object of class \code{hsmm_spec} from the parameters passed as arguments.
}
\examples{
my_model_spec = specify_hsmm(
   J = 2,
   init = c(1,0),
   transition = matrix(c(0,1,1,0),2,2),
   sojourn = list(type = "gamma", shape = c(2,10), scale = c(10,3)),
   parms.emission = list(
       var1 = list(
           type = "norm",
           params = list(
               mean = c(0,1),
               sd = c(0.3,0.2)
               ),
           missing_prob = c(0.6,0.1)
           ),
     var2 = list(
         type = "binom",
         params = list(
             size = rep(1,2),
             prob = c(0.2,0.8)
             ),
         missing_prob = c(0,0)
         ),
     var3 = list(
         type = "non-par",
         params = list(
             values = c("a","b","c","d"),
             probs = matrix(c(0.7,0.1,0.1,0.1,
                            1/4,1/4,1/4,1/4), 4,2)
           ),
         missing_prob = c(0.5,0.5)
         )
     ),
   state_names = c("A","B"),
   state_colors = c("seagreen1","lightcoral")
)
class(my_model_spec)
Xsim = simulate_hsmm(model = my_model_spec)
plot_hsmm_seq(model = my_model_spec, X = Xsim)


}
\keyword{HSMM}
