% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsmm_functions.R
\name{summary_hsmm}
\alias{summary_hsmm}
\title{Prints the model parameters.}
\usage{
summary_hsmm(model)
}
\arguments{
\item{model}{a \code{hsmm} or \code{hsmm_spec} object specifying the hidden semi-Markov model.}
}
\description{
Prints the model parameters.
}
