% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsmm_sojourn_functions.R
\name{gammafit}
\alias{gammafit}
\title{estimates gamma distribution parameters using method of moments}
\usage{
gammafit(x, wt = NULL)
}
\description{
estimates gamma distribution parameters using method of moments
}
