% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_hsmm_joint_em_prob}
\alias{plot_hsmm_joint_em_prob}
\title{Visualizes the joint emission probabilities of a hidden semi-Marvov model.}
\usage{
plot_hsmm_joint_em_prob(model, title = "")
}
\arguments{
\item{model}{a \code{hsmm} object specifying the model for which the joint emission probabilities should be visualized.}

\item{title}{(optional) an character specifying the title of the visualization (typically the name of the model). By default, there is no title.}
}
\value{
a ggplot object.
}
\description{
Visualizes the joint emission probabilities of a hidden semi-Marvov model.
}
