% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_hsmm_transitions}
\alias{plot_hsmm_transitions}
\title{Visualizes the model graph with transition probabilities.}
\usage{
plot_hsmm_transitions(model, layout = NULL)
}
\arguments{
\item{model}{a \code{hsmm} or \code{hsmm_spec} object specifying the model for which the transition probabilities should be visualized.}

\item{layout}{(optional) a matrix of dimension \code{model$J x 2} which provides the coordinates (x,y) of where the states should be placed on the canvas.
If not provided, the layout may slightly differ each time the function is ran.}
}
\value{
a ggplot object.
}
\description{
The edges width is proportional to the transition probability.
}
